refactored and added tests
